modules/person-api/requirements.txt
aniso8601==7.0.0
attrs==19.1.0
Click==7.0
Flask==1.1.1
flask-accepts==0.10.0
flask-cors==3.0.8
Flask-RESTful==0.3.7
flask-restplus==0.12.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.0
GeoAlchemy2==0.8.4
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.0.2
MarkupSafe==1.1.1
marshmallow==3.7.1
marshmallow-sqlalchemy==0.23.1
psycopg2-binary==2.8.5
pyrsistent==0.16.0
python-dateutil==2.8.1
pytz==2020.1
six==1.15.0
shapely==1.7.0
SQLAlchemy==1.3.19
Werkzeug==0.16.1
flask-restx==0.2.0
google
grpcio
protobuf